restructure source layout for some sessions
In case of build failure, the build scripts need to be able to figure out the
AFP entry from the failing session. Since there is no one-to-one mapping from
entry names to session names, we look at the source root of the failing session
and expect this to be a direct subfolder of "thys". Some sessions violated that
expectation. This commit rectifies that.